recreate visualize mesh normals as visualize mesh attributes so that normals and tangents can be both be toggled on or off
